backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-minio-backend>=3.3.1
django-cors-headers>=4.3.0
pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
scipy>=1.7.0
redis>=4.0.0
rq>=1.10.0
gunicorn>=20.1.0
minio>=7.1.0
django-minio-backend
